Heavily reduces contention on generating ids
The main idea is to have batching id generators local to each transaction instance (which is pooled).
This batching sits in between store and typical creators of records at commit time and hands out ids,
grabbing batches of ids off of the stores. On clean shutdown any excess ids remaining unused will be
freed back into the store before closing.

In scenarios where there are many concurrent transactions committing data, where they all
co-operatively force the log then this id generator contention is the single biggest one.
With these changes this contention is basically removed entirely.

The default batch size is 20, which reduces contention dramatically and should be a good value
for most scenarios. The value is however controllable with a database setting.