New cache for in-flight Raft entries
A highly efficient cache which only can keep a consecutive range of raft log entries
cached, allowing for a very simple and efficient circular buffer to be used. The cache
also allows putting a maximum bound on the amount of entries as well as their total combined
size.

The cache also exposes metrics so that cache hits, misses and total data can be observed.