Fix race with channel interceptor
Remove the ChannelInterceptor concept since it potentially left a gap between connect()
and adding the interceptor where messages could slip by without being intercepted. Instead
the complete handshake pipeline is installed immediately from the ChannelInitializer and
the NettyPipelineBuilder has been extended with functionality to cater for the HandshakeGate
which still needs to stay in the pipeline to catch the user event. Such handlers are referred
to as fixed handlers in the builder. I have not been able to create a stress test which shows
the issue, but the issue seems to show up frequently in particular CI tests for loopback
channels.